Created .dockerignore file to optimize Docker build context
This commit adds a .dockerignore file to the repository. The purpose of this file is to optimize the Docker build context by specifying files and directories that should be excluded during the image build process. Excluding unnecessary files like node_modules, logs, and Git-related files helps reduce the size of the build context, resulting in faster and more efficient Docker builds.